feat: converting details card to server component
