+++ class-components/src/components/ButtonDetailedClose/index.ts
@@ -1,0 +1,1 @@
+export { ButtonDetailedClose } from './ButtonDetailedClose';
