+++ class-components/src/components/index.ts
@@ -6,3 +6,4 @@ export * from './SelectedCards';
 export * from './ButtonToHome';
 export * from './ButtonRefreshData';
 export * from './LanguageSelect';
+export * from './ButtonDetailedClose';
